refactor: interceptor로 변경
